% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{get_bin_size}
\alias{get_bin_size}
\title{Get bin size from string resolution}
\usage{
get_bin_size(res)
}
\arguments{
\item{res}{Resolution as a string}
}
\value{
numeric bin size
}
\description{
Get bin size in numeric base pairs from string resolution. Mb, kb, or bp
are supported.
}

