% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{make_is_from_data}
\alias{make_is_from_data}
\title{Make InteractionSet object from a list of data frames}
\usage{
make_is_from_data(raw_df, norm_df, exp_df = NULL, chr, bin_size,
  seqinfo = NULL, interchr = FALSE)
}
\arguments{
\item{raw_df}{data.frame of raw data with bin start positions, sizes, bin numbers, and
observed counts.}

\item{norm_df}{data.frame of normalisation vectors for each bin.}

\item{exp_df}{data.frame of expected vectors, can be NULL to not calculate
expected interactions. Default: NULL.}

\item{chr}{Chromosome of data.}

\item{bin_size}{Numeric bin size in bp.}

\item{seqinfo}{If you don't provide seqinfo, you will get warnings when
combining data from different chromosome. Default: NULL.}
}
\value{
An InteractionSet object with assays containing observed counts and
optionally expected counts. Normalisation vectors are returned as metadata
columns for the `regions()` of the object.
}
\description{
Makes an InteractionSet object from a list of data frames containing raw
counts between bins, normalisation vectors, and (optional) expected data.
}
