% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{normalise_hic}
\alias{normalise_hic}
\title{Normalise observed data by normalisation vectors}
\usage{
normalise_hic(iset, obs = "RAWobserved", norm = NULL)
}
\arguments{
\item{iset}{InteractionSet object with observed interactions as an assay and
normalisation vectors as metadata columns of `regions()`.}

\item{obs}{Name of assay with observed data. Default: 'RAWobserved'.}

\item{norm}{names of columns in `mcols(regions(iset))` to use for normalisation,
or NULL to use all columns matching 'norm'. Default: NULL.}
}
\value{
An InteractionSet object with additional assays holding normalised data.
}
\description{
Normalise observed assays by multiplying them by the appropriate values for
the interacting bins. Result = observed /(anchor 1 value * anchor 2 value)
}
