% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{read_data_from_dir}
\alias{read_data_from_dir}
\title{Read data from given directory into a list of data frames}
\usage{
read_data_from_dir(dir, bin_size, read_expected = FALSE,
  show_progress = TRUE)
}
\arguments{
\item{dir}{Directory with data.}

\item{bin_size}{Numeric bin size}

\item{read_expected}{Whether to read in vectors of expected interactions at
different distances. Default: FALSE.}

\item{show_progress}{Whether to show progress bar for file reading. Default: TRUE.}
}
\value{
A list of data frames, with raw data, normalisation vectors, and
expected data.
}
\description{
Read data from a given directory, for a single chromosome and resolution,
into a list of data frames.
}
