% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{read_data_from_dir_interchr}
\alias{read_data_from_dir_interchr}
\title{Read data from given directory into a list of data frames}
\usage{
read_data_from_dir_interchr(dir, bin_size, show_progress = TRUE)
}
\arguments{
\item{dir}{Directory with data.}

\item{bin_size}{Numeric bin size}

\item{show_progress}{Whether to show progress bar for file reading. Default: TRUE.}
}
\value{
A list of data frames, with raw data, normalisation vectors.
}
\description{
Read data from a given directory, for two different chromosomes and resolution,
into a list of data frames.
}
