% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rao_data_funcs.r
\name{read_rao_huntley_data}
\alias{read_rao_huntley_data}
\title{Read data from a directory}
\usage{
read_rao_huntley_data(dir = ".", resolution = "1Mb", mapq = 30,
  chr = "chr1", inter_chr = FALSE, read_expected = FALSE,
  seqinfo = NULL, show_progress = TRUE)
}
\arguments{
\item{dir}{Data directory containing subdirectories with different resolutions.
Default: current working directory.}

\item{resolution}{Resolution to read in, e.g. 1Mb, 10kb. Default: 1Mb.}

\item{mapq}{MAPQ threshold to use. Only 0 and 30 are supported. Default: 30}

\item{chr}{Chromosomes to read in. Default: chr1.}

\item{inter_chr}{NOT YET IMPLEMENTED. Whether to read in inter-chromosomal
interactions. Default: FALSE.}

\item{read_expected}{Whether to read in vectors of expected interactions at
different distances. Default: FALSE.}

\item{seqinfo}{If you don't provide seqinfo, you will get warnings when
combining data from different chromosome. Default: NULL.}

\item{show_progress}{Whether to show progress bar for file reading. Default: TRUE.}
}
\value{
An InteractionSet object containing data for the specified resolution
and chromosomes.
}
\description{
Read data from the correct subdirectory of a given directory, depending on desired
resolution, mapq, etc.
}

